Backporting to MSVC 11
- Added several new projects to the MSVC 11 solution
- Fixed problems in several examples and headers that
  caused compilation errors with MSVC 11

+++ example/oglplus/010_rgb_cube.cpp
@@ -194,7 +194,7 @@ public:
 		Uniform<Mat4f>(prog, "ProjectionMatrix").Set(
 			CamMatrixf::PerspectiveX(
 				Degrees(48),
-				double(width)/height,
+				GLfloat(width)/height,
 				1, 100
 			)
 		);
